Safari-611.1.10/LayoutTests/http/tests/security/sandboxed-iframe-invalid.html

<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
}
</script>
</head>

<body>
<p>Test that an iframe with invalid sandbox flags generates a relevant warning.</p>

<iframe id="theFrame" sandbox="allowScripts"></iframe>
<iframe id="theFrame2" sandbox="allowScripts allowSameOrigin allowFoobarbloop"></iframe>

</body>
</html>
