Safari-609.1.11.1/LayoutTests/http/tests/security/sandboxed-iframe-invalid-expected.txt

CONSOLE MESSAGE: line 13: Error while parsing the 'sandbox' attribute: 'allowScripts' is an invalid sandbox flag.
CONSOLE MESSAGE: line 14: Error while parsing the 'sandbox' attribute: 'allowScripts', 'allowSameOrigin', 'allowFoobarbloop' are invalid sandbox flags.
Test that an iframe with invalid sandbox flags generates a relevant warning.

 
